Refactor dataset groupby tests
