+++ xarray/tests/__init__.py
@@ -6,11 +6,13 @@ from distutils import version
 from unittest import mock  # noqa: F401
 
 import numpy as np
+import pandas as pd
 import pytest
 from numpy.testing import assert_array_equal  # noqa: F401
 from pandas.testing import assert_frame_equal  # noqa: F401
 
 import xarray.testing
+from xarray import Dataset
 from xarray.core import utils
 from xarray.core.duck_array_ops import allclose_or_equiv  # noqa: F401
 from xarray.core.indexing import ExplicitlyIndexed
@@ -200,3 +202,30 @@ def assert_allclose(a, b, **kwargs):
     xarray.testing.assert_allclose(a, b, **kwargs)
     xarray.testing._assert_internal_invariants(a)
     xarray.testing._assert_internal_invariants(b)
+
+
+def create_test_data(seed=None, add_attrs=True):
+    rs = np.random.RandomState(seed)
+    _vars = {
+        "var1": ["dim1", "dim2"],
+        "var2": ["dim1", "dim2"],
+        "var3": ["dim3", "dim1"],
+    }
+    _dims = {"dim1": 8, "dim2": 9, "dim3": 10}
+
+    obj = Dataset()
+    obj["dim2"] = ("dim2", 0.5 * np.arange(_dims["dim2"]))
+    obj["dim3"] = ("dim3", list("abcdefghij"))
+    obj["time"] = ("time", pd.date_range("2000-01-01", periods=20))
+    for v, dims in sorted(_vars.items()):
+        data = rs.normal(size=tuple(_dims[d] for d in dims))
+        obj[v] = (dims, data)
+        if add_attrs:
+            obj[v].attrs = {"foo": "variable"}
+    obj.coords["numbers"] = (
+        "dim3",
+        np.array([0, 1, 2, 0, 0, 1, 1, 2, 2, 3], dtype="int64"),
+    )
+    obj.encoding = {"foo": "bar"}
+    assert all(obj.data.flags.writeable for obj in obj.variables.values())
+    return obj
